feat: add submit and wait for transaction submission (#528)
Currently, user has to call 2 functions (autofill_and_sign and send_reliable_submission) in order to sign, autofill, submit and verifies that the transaction has been included in a validated ledger. This PR would simplify the workflow to a single call (submit_and_wait) similar to xrpl.js (the old workflow would still work so it is not a breaking change). Tests and snippets are also fixed to use the new method.

Co-authored-by: Mayukha Vadari <mvadari@gmail.com>
Co-authored-by: Jackson Mills <aim4math@gmail.com>